+++ src/main/java/Lesson_5/Fruit/Pear.java
@@ -1,20 +1,22 @@
 package Lesson_5.Fruit;
 
 public class Pear extends Fruit {
-    private static int totalPrice;
+    private static double totalPrice;
 
-    public Pear(int weight, int price) {
+    public Pear(double weight, double price) {
         super(weight, price);
         totalPrice += price();
     }
 
-    public static int totalPearPrice() {
+    public static double totalPearPrice() {
         return totalPrice;
     }
 
     @Override
-    public int price() {
+    public double price() {
         return getPrice() * getWeight();
     }
+
+
 
 }
